test can parsing
